+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1254
-    const val appVersion = "4.21.8.2"
+    const val appCodeNumber = 1255
+    const val appVersion = "4.21.8.3"
     const val identifier = "com.unciv.app"
 }